Add flag for percentage writes
This commit adds a flag for the percentage of writes we want in the
run.

To do this we add separate read and write work queues. We add the
DoWrite type which lets the main run loop know whether a read/write
is needed for this iteration, using a simple 100 item lookup table.
DoWrite's Next() function is stable for given percentages to allow
for recreating runs.

Finally we add a summary of the read/writer results for easier reading
(ie, easier to check the percentage looks right :D )